Sistemata autenticazione e bug registrazione senza avatar

+++ src/main/java/it/unitn/webprog2018/ueb/shoppinglist/servlets/auth/AccountConfirmationServlet.java
@@ -33,10 +33,10 @@ public class AccountConfirmationServlet extends HttpServlet {
 
 	private UserDAO userDAO;
 	private TokenDAO tokenDAO;
-	
+
 	/**
-	 * Method to be executed at servlet initialization.
-	 * Handles connections with persistence layer.
+	 * Method to be executed at servlet initialization. Handles connections with
+	 * persistence layer.
 	 */
 	@Override
 	public void init() {
@@ -65,13 +65,17 @@ public class AccountConfirmationServlet extends HttpServlet {
 		String tokenString = request.getParameter("token");
 
 		Token token = tokenDAO.getByTokenString(tokenString);
-		
+		String avatarName = "";
+		String uploadFolder = getServletContext().getInitParameter("uploadFolder");
 		if (token == null) {
 			// TODO edit to correct page
 			path += "invalidToken.html";
 			response.sendRedirect(path);
 		} else if (isExpired(token)) {
 			tokenDAO.removeToken(token);
+			avatarName = token.getUser().getImage().substring(token.getUser().getImage().lastIndexOf("/") + 1);
+			File avatar = new File(uploadFolder + "/restricted/tmp/" + avatarName);
+			avatar.delete();
 			// TODO redirect to error page
 			path += "expiredToken.html";
 			response.sendRedirect(path);
@@ -79,23 +83,21 @@ public class AccountConfirmationServlet extends HttpServlet {
 		if (!response.isCommitted()) {
 			User user = token.getUser();
 			try {
-				if(userDAO.addUser(user))
-				{
-					String avatarName = user.getImage().substring(user.getImage().lastIndexOf("/") + 1);
-			
-					String uploadFolder = getServletContext().getInitParameter("uploadFolder");
-					File src = new File(uploadFolder + "/restricted/tmp/" + avatarName);
-					String avatarName2 = avatarName.replaceFirst(user.getEmail(), user.getId().toString());
-					File dest = new File(uploadFolder + "/restricted/avatar/" + avatarName2);
-					Files.copy(src.toPath(), dest.toPath());
-					src.delete();
+				user.setCheckpassword(user.getPassword());
+				if (userDAO.addUser(user)) {
+					if (!user.getImage().equals("")) {
+						avatarName = user.getImage().substring(user.getImage().lastIndexOf("/") + 1);
 
+						File src = new File(uploadFolder + "/restricted/tmp/" + avatarName);
+						String avatarName2 = avatarName.replaceFirst(user.getEmail(), user.getId().toString());
+						File dest = new File(uploadFolder + "/restricted/avatar/" + avatarName2);
+						Files.copy(src.toPath(), dest.toPath());
+						src.delete();
+					}
 					tokenDAO.removeToken(token);
 					path += "Login";
 					response.sendRedirect(path);
-				}
-				else
-				{
+				} else {
 					request.setAttribute("user", user);
 					request.getRequestDispatcher("/WEB-INF/views/auth/SignUp.jsp").forward(request, response);
 				}
